add fuzzy key to member ID mapping
